Renamed store to configureStore
