+++ resources/react/configureStore.js
@@ -1,9 +1,9 @@
+import 'isomorphic-fetch';
 import { createStore, applyMiddleware, compose } from 'redux';
 import { routerMiddleware } from 'react-router-redux';
 import thunk from 'redux-thunk';
 import promiseMiddleware from './middleware/promiseMiddleware';
 import createReducer from './createReducer';
-import createLogger from 'redux-logger';
 
 export function configureStore (initialState, history) {
     const middleware = [
@@ -15,6 +15,8 @@ export function configureStore (initialState, history) {
     let store;
 
     if (__DEV__ && __CLIENT__) {
+        const createLogger = require('redux-logger');
+
         middleware.push(createLogger());
         const reducer = createReducer();
 
@@ -29,10 +31,12 @@ export function configureStore (initialState, history) {
 
     store.asyncReducers = {}
 
-    if (__DEV__) {
-        if (module.hot) {
-            module.hot.accept('./createReducer', () => store.replaceReducer(require('./createReducer').default));
-        }
+    if (__DEV__ && __CLIENT__ && module.hot) {
+        module.hot.accept('./createReducer', () => {
+            const nextReducer = require('./createReducer').default;
+
+            store.replaceReducer(nextReducer());
+        });
     }
 
     return store;
